Fix filename cleansing within "create_fileindex"
The description of the "lstrip" method is
"Return a copy of the string with leading characters removed."

While that sounds nice, it is not quite what we want here.

The syntax is actually "str.lstrip([chars])" where "chars" is
a string specifying the **set of characters** to be removed.

The chars argument is not a prefix; rather, all combinations 
of its values are stripped.

Actually, this goes south when trying to "strip away" the
DWD_PATH prefix because "annual" will become "ual".