Compensate compensation for DWD folder anomaly at "hourly solar"
The background of this is that the DWD folder
https://opendata.dwd.de/climate_environment/CDC/observations_germany/climate/hourly/solar/ does not discriminate between historical and
recent data.

We figure that it might be more convenient for the user to accept
any PeriodType value and thus we will not even have to document
this anomaly.

The implementation under the hood already compensates for this anomaly
and so we should give this feature thoroughly back to the user without
sacrificing it by pushing the obstacle to another place.