Fix Settings once again
Using dataclass as core class for our Settings container previously led to multiple problems in combination with a property we used for logging. This patch removes the dataclass dependency and uses simple init.

Furthermore, the logging level is changed to INFO as was reported that the warning level is cloaking the logging system too much.