Make explorer work again
The explorer previously didn't work for some weather services which had a fixed period and for which we set an instance as period base rather than an enum with a single value. Now with an enum class set as period base they work once again.

Also for ECCC Observation the date range was previously set to loosely which would require setting start date and end date. However the data provided actually is given in chunked files and setting was changed to date range fixed. This now also works once again.